+++ MaterialColorUtilities/Quantize/PointProviderLab.cs
@@ -4,22 +4,22 @@ namespace MaterialColorUtilities.Quantize;
 
 public class PointProviderLab : IPointProvider
 {
-    public float[] FromInt(int argb)
+    public double[] FromInt(int argb)
     {
         double[] lab = ColorUtils.LabFromArgb(argb);
-        return new float[] { (float)lab[0], (float)lab[1], (float)lab[2] };
+        return new double[] { lab[0], lab[1], lab[2] };
     }
 
-    public int ToInt(float[] lab)
+    public int ToInt(double[] lab)
     {
         return ColorUtils.ArgbFromLab(lab[0], lab[1], lab[2]);
     }
 
-    public float Distance(float[] one, float[] two)
+    public double Distance(double[] one, double[] two)
     {
-        float dL = one[0] - two[0];
-        float dA = one[1] - two[1];
-        float dB = one[2] - two[2];
+        double dL = one[0] - two[0];
+        double dA = one[1] - two[1];
+        double dB = one[2] - two[2];
         return dL * dL + dA * dA + dB * dB;
     }
 }
